requirements.txt
aniso8601==1.1.0
click==6.6
contextlib2==0.5.4
cryptography==1.5
enum34==1.1.6
Flask==0.11.1
Flask-API==0.6.7
Flask-Cors==2.1.2
Flask-Fixtures==0.3.7
Flask-JWT==0.3.2
flask-marshmallow==0.7.0
Flask-Migrate==2.0.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.12
six==1.10.0
sqlalchemy-utils==0.32.9
Werkzeug==0.11.10
MySQL-python==1.2.5
PyJWT==1.4.1
SQLAlchemy==1.0.14
validate-email==1.3
Mako==1.0.3
requests==2.10.0
Flask-MongoAlchemy==0.7.2
boto==2.43.0
boto3==1.4.0
botocore==1.4.70
marshmallow==2.10.3

